Fix 64-bit pointer truncation in sprite_getFramePtr
The (s32) cast truncated a 64-bit pointer to 32 bits, producing an
invalid address when casting back to BKSpriteFrame*. This caused
SEGV_MAPERR at 0x40465c76 in print_getLettersFromFont.

Removed the cast to preserve full 64-bit pointer arithmetic.

+++ src/core1/sprite.c
@@ -23,6 +23,6 @@ s32 sprite_getFrameCount(BKSprite *this) {
 
 BKSpriteFrame *sprite_getFramePtr(BKSprite *this, u32 frame_id) {
     BKSpriteFrame *sprite_frame;
-    sprite_frame = (BKSpriteFrame *) (this->offsets[frame_id] + ((s32) ((*this).offsets + this->frameCnt)));
+    sprite_frame = (BKSpriteFrame *) (this->offsets[frame_id] + (((*this).offsets + this->frameCnt)));
     return sprite_frame;
 }
